ref(ui): Move bulkController + bulkNotice into own folder inside components
