+++ tests/js/spec/components/bulkController/index.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import {shallow} from 'sentry-test/enzyme';
 
-import BulkController from 'app/utils/bulkController';
+import BulkController from 'app/components/bulkController';
 
 describe('BulkController', function() {
   let wrapper,
